1. change blas_interface to blas_connector 	1.1 src_global/blas_connector.h 	1.2 src_global/lapack_connector.h 	1.3 src_external/src_pdiag/blas_interface.h 	1.4 src_external/src_pdiag/pdiag_common.h 	1.5 src_external/src_pdiag/pdiag_basic.cpp 	1.6 src_external/src_pdiag/pdiag_double.cpp 	1.7 src_external/src_pdiag/pdstebz.cpp 	1.8 src_external/src_pdiag/pdsytrd.cpp 	1.9 src_external/src_pdiag/pzgseps.cpp 	1.10 src_external/src_pdiag/pzhetrd.cpp 	1.11 src_lcao/evolve_lcao_matrix.cpp 	1.12 src_lcao/gint_gamma_fvl.cpp 	1.13 src_lcao/gint_gamma_rho.cpp 	1.14 src_lcao/gint_gamma_vl.cpp 	1.15 src_lcao/gint_k_rho.cpp 	1.16 src_lcao/gint_k_vl.cpp 	1.17 src_lcao/local_orbital_charge.cpp 	1.18 src_pw/hamilt_pw.cpp 	1.19 src_pw/myfunc5.cpp
2. fix range bug, change some int to long
	2.1 src_global/global_function.h
	2.2 src_external/src_pdiag/pdiag_double.h
	2.3 src_external/src_pdiag/pdiag_double.cpp
	2.4 src_lcao/lcao_matrix.h
	2.5 src_lcao/lcao_matrix.cpp
	2.6 src_global/memory.h
	2.7 src_global/memory.cpp

3. move Charge::new_e_iteration to private
	3.1 src_pw/charge_pulay.h
	3.2 src_pw/charge_pulay.cpp
	3.3 src_external/src_pdiag/pdiag_double.cpp
	3.4 src_lcao/gint_gamma_vl.cpp
	3.5 src_lcao/local_orbital_elec.cpp
	3.6 src_pw/electrons.cpp

+++ ABACUS.1.0.0/source/src_lcao/lcao_matrix.h
@@ -143,8 +143,8 @@ class LCAO_Matrix
 	void update_Hloc2(void);
 
 	void allocate_HS_R(const int &nnr);
-	void allocate_HS_gamma(const int &nloc);
-	void allocate_HS_k(const int &nloc);
+	void allocate_HS_gamma(const long &nloc);
+	void allocate_HS_k(const long &nloc);
 
 	void output_HSk(const char &mtype, string &fn);
 	//LiuXh add 2019-07-15
